+++ spec/rubocop/cop/rails/redundant_receiver_in_with_options_spec.rb
@@ -37,6 +37,27 @@ describe RuboCop::Cop::Rails::RedundantReceiverInWithOptions, :config do
       RUBY
     end
 
+    it 'registers an offense when including multiple redendant receivers ' \
+       'in single line' do
+      expect_offense(<<-RUBY.strip_indent)
+        with_options options: false do |merger|
+          merger.invoke(merger.something)
+          ^^^^^^ Redundant receiver in `with_options`.
+                        ^^^^^^ Redundant receiver in `with_options`.
+        end
+      RUBY
+    end
+
+    it 'does not register an offense when including method invocations ' \
+       'to different receivers' do
+      expect_no_offenses(<<-RUBY.strip_indent)
+        client = ApplicationClient.new
+        with_options options: false do |merger|
+          client.invoke(merger.something, something)
+        end
+      RUBY
+    end
+
     it 'autocorrects to implicit receiver in `with_options`' do
       new_source = autocorrect_source(<<-RUBY.strip_indent)
         class Account < ApplicationRecord
@@ -56,6 +77,32 @@ describe RuboCop::Cop::Rails::RedundantReceiverInWithOptions, :config do
             has_many :products
             has_many :invoices
             has_many :expenses
+          end
+        end
+      RUBY
+    end
+
+    it 'autocorrects to implicit receiver when including multiple receivers' do
+      new_source = autocorrect_source(<<-RUBY.strip_indent)
+        with_options options: false do |merger|
+          merger.invoke(merger.something)
+        end
+      RUBY
+
+      expect(new_source).to eq(<<-RUBY.strip_indent)
+        with_options options: false do
+          invoke(something)
+        end
+      RUBY
+    end
+
+    it 'does not register an offense when including block node' \
+       'in `with_options`' do
+      expect_no_offenses(<<-RUBY.strip_indent)
+        with_options options: false do |merger|
+          merger.invoke
+          with_another_method do |another_receiver|
+            merger.invoke(another_receiver)
           end
         end
       RUBY